releases/describe_firmware: remove bitbox02 dependency
As per the releases/README, the file is supposed to be easy to
audit. Pulling in a dependency makes audit more difficult. It also
makes the script harder to use, if the user does not have the
dependency installed.

This commit inlines the relevant parse_signed_firmware() function from
bootloader.py.