ci: lint all default python files to avoid errors
If e.g. only send_message.py is linted without the bitbox02 package
files, import errors can happen.

We still only start linting if any relevant Python file has been
touched, but then we lint all of them.

The reason if sometimes worked before is that the bitbox02 package
installed in the system was used (installed in Docker during Docker
build), but that was linting the wrong package.