cardano: add validity_interval_start
Since the Allegra update, the TTL transaction field has been replaced
with the validity interval `[start, end]`. See
https://hydra.iohk.io/build/7840126/download/1/shelley-ma.pdf for
details.

All wallets I looked at only use the validity interval start. The cddl
specification file also is explicit about the validity interval start,
but not about the end:

https://github.com/input-output-hk/cardano-ledger-specs/blob/d0aa86ded0b973b09b629e5aa62aa1e71364d088/eras/alonzo/test-suite/cddl-files/alonzo.cddl#L59

It seems to me that the validity interval end and TTL are the same,
and that the TTL field is used for that purpose since Allegra.