cardano: verify ttl and validity_interval_start
Instead of showing the absolute slot numbers, we encode them as the
relative slot number to an epoch. This should be easier to interpret
for a user, as an epoch lasts about 5 days on mainnet, and the current
epoch can be checked on block explorers or full nodes.

I could not find docs about the epoch params in the official Cardano
spec at https://github.com/input-output-hk/cardano-ledger-specs.

I gathered them from:

- https://cardano-foundation.gitbook.io/stake-pool-course/lessons/introduction/about-cardano (432000
slots per epoch)

- https://github.com/LedgerHQ/app-cardano/blob/c57e8410edd2185062fb6493720fd1533365877e/src/textUtils.c#L173