cardano: don't show total if it is 0
When registering a staking key, delegating or withdrawing, typically
there are no outgoing funds except for the fee.

The final screen would show both total and fee with the same
amount. In this case, we now only show the fee, as there is no
relevant total.

It also helps avoid confusion with wallets such as Adalite, where the
total can be the total outgoing or the total incoming sum depending on
the operation. If withdrawing staking rewards for example, Adalite
would show the total of funds incoming, which is the rewards minus the
fees, whereas the BitBox02 only shows outgoing amounts when confirming
a total.