btc: rename output `hash` to `payload`
When confirming a tx, the recipient address is reconstructed from the
payload provided in the protobuf `BTCSignOutputRequest` type and
displayed for confirmation.

We previously named the payload `hash` as it always was a hash: either
a pubkey hash in case of P2WPKH (pay to witness pubkey hash), or a
script hash in case of P2SH (pay to scripthash) or P2WSH (pay to
witness script hash).

With the Taproot softfork that recently activated on Bitcoin, the
payload is no longer a hash of something but a 32 byte public key. See
https://github.com/bitcoin/bips/blob/master/bip-0341.mediawiki ("Why
is the public key directly included in the output?").

"Payload" as a term is also used in other Bitcoin libraries for the
decoded address input, e.g. in rust-bitcoin. We apply the same rename
to the relevant helper functions in btc_common.h.

This is not a breaking change for the BitBox02 API, as the protobuf
names are not present in the binary serialization, so existing
downstream integrations are not affected.

Projects that update to the changed protobuf file will have to rename
`hash` to `payload` in the code generated from the protobuf files.

Alternatives considered:

- Keep `hash` as is, add another field `pubkey`. This puts as much
  burden on the client libraries as renaming to payload, but adds
  complexity to the firmware code.
- Add `encoded_address`, and let the firmware decode it into the
  payload and validate it. This would be a new way of showing
  recipient addresses, but we'd need to keep supporting the old way
  for backwards compatibility. In some inputs, e.g. PSBT, there is no
  address string to provide, so it does not even make it easier for
  downstream projects in all cases.

It seems that the rename was the easiest solution.