py/bitbox02: fix enum types
In an older version of mypy-protobuf, enums were type-ignored.

Now they work, but we have to use the `.V` type.

They are marked as strings as they only exist at typechecking time,
not runtime. In an upcoming version of mypy-protobuf, we will be able
to use `.ValueType` instead without the string quotes.

See also:

https://github.com/protocolbuffers/protobuf/issues/8175

https://github.com/protocolbuffers/protobuf/commit/0707f2e7f556c8396d6027d0533ec3a56d1061db