py/bitbox02: fix imports in protobuf pyi type stub files
E.g. in hww_pb2.pyi, the import to `bitbox02_system_pb2` could not be
found as the folder is not in any mypy or python PATH. As a result,
protobuf message field accesses were not properly type checked before.

There is already a line in Makefile to fix this, but it did not
work (or does not work anymore). We have to turn the imports into
relative imports to get mypy to find and use the types.