scripts/lint-python: fix invocation in updated mypy
Mypy now is severely confused by there being a bitbox02.py inside a
bitbox02 folder, resulting in duplicate module errors. See
https://github.com/python/mypy/issues/4008.

gvanrossum writes:

> This is because apparently you seem to have both a folder named gsi
and a file name gsi.py. This is a problem for Python too (IIRC the
folder/package wins) so you can't really blame mypy for complaining
about this.

I don't understand this, as Python seems to handle it fine in our
package, but mypy does not.