cardano: allow ttl=0
Before, ttl=0 was interpreted as "not provided", which means that it
is not serialized and signed.

According to the spec, zero is a valid value for the TTL. It means
however that the transaction can never be mined, as the current slot
will always be bigger than 0.

According to the Adalite team, this type of transaction is used in
meta protocols:
https://github.com/vacuumlabs/adalite/pull/1212#discussion_r755973295

> But TTL = 0 is in theory a semantically valid TTL the transaction
can have - advanced users may leverage that to sign a transaction and
ensure it won't ever be accepted by the blockchain - it's used
e.g. with Ledger when CLI users need to sign separately voting
registration metadata (but the device allows signing only the full tx)
and therefore want to throw away the rest of the signed
transaction. (of course, they can use ttl=1 which would do as well,
but mapping 0 to null seems worth avoiding for the sake of
predictability of the interface and overall consistency)

To be backwards compatible, we add a new protobuf field
`allow_zero_ttl` and take it into account. "Transaction cannot be
mined" is shown to the user in this case.

Since it touches the same code, I also extended the "Transaction
cannot be mined" case to when the validity start is bigger than the
TTL.

A note on the new protobuf field:
- it is a new field that does not change the behvior if set to the
  default, false. This is for compatibility for older client libraries
  talking to newer firmware.
- `optional` would have been great, and is available in a newer proto3
  version, but prost does not support it yet, and possibly it is also
  not supported by other protobuf code generators.