keystore: add keystore_secp256k1_schnorr_bip86_[pubkey|sign] funcs
Needed to to generate taproot receive addresses and sign taproot
inputs in transactions.

A taproot output contains a 32 byte x-only pubkey (see BIP-340), which
may or may not be tweaked with the root of a merkle tree containing
execution scripts as leafs.

In a normal single-sig with no extra scripts, BIP-86 specifies:

> BIP 341 states: "If the spending conditions do not require a script
path, the output key should commit to an unspendable script path
instead of having no script path. This can be achieved by computing
the output key point as Q = P + int(hashTapTweak(bytes(P)))G.".

https://github.com/bitcoin/bips/blob/4c6389f8431f677847b115538a47ce8c826c6be8/bip-0086.mediawiki

This commit adds keystore functions to get such tweaked
pubkey (original pubkey tweaked with the hash of the original pubkey)
and to sign for it.

We put this functionality directly into keystore.c instead of
performing the tweaking outside of it so we only have to derive the
pubkey at the keypath once. The alternative would have been to add a
32 bytes tweak argument to the pubkey/sign functions and have the
caller compute the correct tweak, or to perform the tweaking at the
callsite altogether.