app/btc/sign: move btc tx sign state machine to Rust async funcs
The Bitcoin signing algorithm in btc_sign.c involves keeping track of
a big state machine, processing the transaction in a streaming
fashion.

The `next_request()` async primitive is perfect to stream the
transaction from the host in an async function, offloading the state
machine to the Rust compiler, improving readability.

Instead of migrating the whole implementation to Rust in one shot, at
first we only move streaming the transaction to Rust, and call the
existing signing functions of C. `_wrapper()` functions are added for
each of them so they can be called from Rust.

The C implementation is largely unchanged in this commit. In the
future, we can get rid of all the state handling in the C
implementation.

In btc_sign.c, the only changed needed was make _sign_input_pass1()
and _sign_input_pass2() work indepently without relying on
`app_btc_sign_input(). `app_btc_sign_input()` is now only used in C,
while in Rust the two passes are called individually.

Over time, these functions can move to Rust as well.

To ensure that everything works, we port some tests from
test_btc_sign.c:

- _test_btc_sign_init()
- `_sign()` with the default `_modification_t`, so the unmodified
  transaction. We compare the produced signature at the end, which is
  a copy paste from test_btc_sign.c.