app/btc/sign: move progress bar and empty screen to Rust
The progress bar and empty screen depend on the state machine in
btc_sign.c. To get rid of the state machine in C, we first need to
move out this logic out of there.

In the async Rust function, updating the progress bar and
pushing/popping the empty screen is much easier to understand.