rust: run unit tests serially, drop MUTEX to manually serialize
Before, if a unit test panicked with a mutex guard, all other unit
tests waiting for the mutex panic as well due to being poisoned. This
leads to very verbose and messy output if a unit tests
fails. Furthermore, the actual unit test error is sometimes mangled up
with the other output.

By running the tests serially, we can drop the mutex and avoid all
these problems.

Alternative considered: use into_inner() on the PoisonError, as
described in

- https://doc.rust-lang.org/std/sync/struct.PoisonError.html
- https://doc.rust-lang.org/nomicon/poisoning.html

It is not clear if this would always lead to correct unit test
behavior, and handling the locks becomes even less ergonomic.