api/eth: use chain ID to identify the network, add BNB/MATIC
Until now a BitBox2 client had to provide the network they want to
interact using the ETHCoin enum. This has the downside that the list
of supported networks has to be mirrored in the client library, to map
from the chain id (usually provided by an ETH wallet such as
MyEtherWallet/Rabby) to the ETHCoin enum.

To avoid this duplication and make things simpler, we now allow the
client to provide the chain ID instead. ETHCoin is kept for backwards
compatibility.

BSC is added as a first network that can only by identified by chain
ID.