eth: allow unknown networks after confirming chain ID and warning
If a user receives funds on a EVM-compatible network not supported
natively by the firmware, after this commit a user can still sign a
transaction after confirming the chain ID and a warning. The purpose
is mainly to allow users to somewhat safely recover funds which would
otherwise be stuck.