memory: move shared data to new file memory_shared.h
The bootloader and firmware both access the same memory starting at
FLASH_SHARED_DATA_START. The bootloader doesn't link memory.c, and had
a local copy of the relevant struct.

This creates a memory_shared.c linked into both bootloader and
firmware for better code reuse. This is done in preparation of adding
a `screen_type` memory field which will determine which screen driver
to use.