unit-test: less manual mocking of securechip_kdf
We don't need to mock the input and output of the kdf function in each
unit test - by default it should do the same as the actual securechip,
but with mocked slot contents. This allows unit tests that run
securechip_kdf to work in Rust as well, where we cannot (or will not!)
use cmocka mocking.