rust: define rust_eh_personality
When updating to Rust 1.61, linking the C unit tests fails with

```
../../lib/libc-unit-tests_rust_c.a(bitbox02_rust_c-d1aed3e812268234.bitbox02_rust_c.2b72ae46-cgu.0.rcgu.o):(.data.DW.ref.rust_eh_personality[DW.ref.rust_eh_personality]+0x0):
undefined reference to `rust_eh_personality'
```

This section explains that this is needed when writing an executable:

https://doc.rust-lang.org/unstable-book/language-features/lang-items.html#writing-an-executable-without-stdlib

It is not clear why it was not needed before updating Rust.

The `dont-export-eh-personality` feature is needed to exclude it when
compiling Rust unit tests, as some of them depend on `std`.

Co-authored-by: Niklas Dusenlund <niklas.dusenlund@safello.com>