eth/signtypedmsg: show progress in title
As typed data is streamed from the host, the total number of
confirmation pages is unknown upfront. We do know however the number
of members in the top level struct. We show the current step and total
for the top level struct, but for any given step, there can be lots of
confirmation pages if the member another struct or array.