releases/build.sh: fix reproducible builds that failed due to git
A breaking change happened in git which disallows git commands in a
repo owned by someone else:

```
ailed to run `['git', 'tag', '--list', '--sort=taggerdate', '--points-at', 'HEAD', '--format=%(objecttype) %(if)%(contents:signature)%(then)Y%(else)N%(end) %(refname:strip=3)', 'firmware-btc-only/*']`
stderr: fatal: unsafe repository ('/bb02' is owned by someone else)
To add an exception for this directory, call:

	git config --global --add safe.directory /bb02
```

Related to this fix:

https://github.blog/2022-04-12-git-security-vulnerability-announced/

We need git commands to get the firmware version, which is part of the
binary. With the error above, the version wat not included in the
binary properly, resuling in a wrong build.

We implement the fix suggested in the error msg. The alternative is
run the docker image with the same UID as the host UID, but I am not
sure if this works as well on all host systems.