backup: initialize backup stack vars to zero
backup_check and backup_create compute the directoy name of the backup
based on the seed field in the backup struct, using
`_get_directory_name`, always using 32 bytes, even if the seed is
shorter (e.g. 16 bytes).

Without explicitly zero-intializing these structs and the seed is
shorter than 32 bytes, it could contain garbage instead of zero as
padding. If that happened, backup creation and restore would work, but
the backup check could fail, as the directory could not be
found (directory name computed based on a seed with garbage padding).

Somehow it seems that the structs are zero initialized in
practice (confirmed by manual testing), but that seems to be a lucky
coincidence. This commit makes this explicit.