rust/sd: fix list_subdir ffi call
The subdir argument was not passed to C properly - the result
of str_to_cstr_vec was dropped before `bitbox02_sys::sd_list_subdir()`
is called.

This commit fixes it by making sure the result is retained until after
the call is finished.

This function was not used yet with the subdir arg, so there was no
bug resulting from this.