backup/checksum: make size of BackupMode explicit
In C, the size of an enum is undefined. In my manual checks,
`sizeof(BackupMode)` is 4 bytes on my host machine and 1 byte on the
BitBox02.

We explicitly set the size of the field for clarity to 1
byte (compatible with all backups created on the BitBox02s).

The unit test fixtures needed to change because it was computed using
the size of the field on the host (4 bytes instead of 1 byte).