backup: unit test against backup fixtures
We want to port the backup code to Rust and also possibly do some
modifications, without worrying that we will break backups made by
older firmware versions.

This commit adds fixtures from a backup created using firmware v9.12.0
and checks that the backup is valid and contains the right
seed. Hopefully this will test regressions when working on the backup
code.