prost/protobuf: generate backup.proto for Rust
Make the backup.proto protobuf messages available in Rust, so we can
work with microSD backups in Rust.

The package name of backup.proto is changed as it is 100% independent
of the other .proto files, which all deal with the firmware API. This
results in the different file for the definitions in Rust.

The Docker image is updated because prost-build as modified and it is
pre-compiled during image generation.