eth/params: remove long name from erc20 tokens
This saves 33456 bytes of firmware space.

The full names takes up a lot of firmware space.

The long name is currently only used when displaying a receive address
in the BitBoxApp. In this commit, we use the unit instead of the long
name there.

I added the coin/network name to the title as well, just in case it
becomes helpful (e.g. if we end up having the same token on multiple
EVM chains).