i2c_ecc,random: remove duplicate definitions
These symbols are already defined in driver_init.c.

The current GCC version doesn't complain, I discovered this when
trying to build with an updated
GCC (arm-gnu-toolchain-11.3.rel1-x86_64) and getting errors about
these redefinitions.

The reason that the current GCC version (GCC 8) didn't complain and
the updated one (GCC 11) did complain is that the default for the
-fcommon flag changed from on to off in GCC 10.

See:
- https://gcc.gnu.org/bugzilla/show_bug.cgi?id=85678
- https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html (-fcommon)

clang-tidy also checked i2c_ecc.c because it was modified and reported
these errors, which are also fixed in the commit:

```
/bitbox02-firmware/src/i2c_ecc.c:29:18: error: narrowing conversion from 'uint32_t' (aka 'unsigned int') to signed type 'int32_t' (aka 'int') is implementation-defined [bugprone-narrowing-conversions,-warnings-as-errors]
    packet.len = rxlen;
                 ^
/bitbox02-firmware/src/i2c_ecc.c:48:18: error: narrowing conversion from 'uint32_t' (aka 'unsigned int') to signed type 'int32_t' (aka 'int') is implementation-defined [bugprone-narrowing-conversions,-warnings-as-errors]
    packet.len = txlen;
```