api/setpassword,restore: return UserAbort on user abort
Before, Generic would be returned no matter if the user did not enter
the same password twice or if they aborted on the `Your password has
fewer than 4 characters.\nContinue?` screen.

Now that we handle the user abort, we can also allow to abort during
password entry itself.

We also add cancel confirmations "Do you really want to cancel?" when
cancelling the password screen or the password-too-short-screen, as it
would be highly frustrating to accidentally cancel after entering a 24
word mnemonic or a long password.