api: display error type on screen if create/restore fails
If there was a securechip or memory error, creating or restoring a
wallet would show and return a generic looking error. If e.g. the
securechip is the problem, we want to see that, so the user can better
report the problem.

We might also return a specific error code to the host, but for now
displaying the error will do (same as do when unlocking a device fails).