releases/build.sh: fix to work with recently updated Docker image
In 11eea89dffe215461327f8e7a34fbf8cd75ffa07, we updated the base image
to Ubuntu 22.04.

`python` is not an installation candidate any longer. We simply remove
it, as it was a fix for very old versions, and keep a comment in case
someone wants to build these old versoins.

Somehow it seems that `git clone --depth 1` doesn't fetch the tags
anymore - fixed by fetching the tags manually.