rust: update rust toolchain to 1.66.0 and cbindgen to 0.63.0
cbindgen made a breaking change renaming whitelist to allowlist.

The rest is new Clippy warnings.

Clippy needless_borrow is disabled because the new cases it reported
seemed to cause a move/copy. I am not sure if the compiler would have
created the same output in either case. Might revisit this and
re-enable this linter in the future.