keystore: use xpubs instead of xprvs where possible
Functionally the same, but for clarity it is better to use xpubs where
only pubkeys are involved. This will also make it easier to add xpub
caching to increase performance if needed (for example, xpubs are
derived anew for each BTC transaction input and change).

The extra keystore_is_locked check is not needed getting a xprv/xpub
already does that check.