tools/atecc608a: update doc
The unsafeSignKeySlot has been in use to sign U2F things for a long
time, so it is not reserved for future use as the doc says, but used
today.

We remove meniton of which are in use and which are not. securechip.h
contains a list of actively used slots.