keystore: copy instead of returning static pointer
Currently the seed is in RAM in plaintext after unlock. We want to
encrypt it instead, using the secure chip.

This refactors the _get_seed and _get_bip39_seed functions to not
simply return the static pointer, but a copy, allowing us to perform
decryption later once the seeds are encrypted in RAM.