unit-test/keystore: simplify by removing some mocks
These unit tests are hard to maintain as there is a lot of mocking
with will_return and expect_*. We don't need to mock salt_hash_data
and securechip_kdf results, we use existing mock implementations of
these that perform the real function (in case of securechip_kdf, the
mocked one from mock_securechip.c).

For securechip_kdf we dont need to check each input as the secret is
checked using _expected_secret, which is the result of the whole key
stretching (including securechip_kdf).