rust: remove str_to_cstr and str_to_cstr_force
Simplify code by using str_to_cstr_vec instead (which we already use
in many places for the same purpose), which uses core::alloc::CString
for conversion. Since it uses a Vec (heap), one does not have to
determine a mac size upfront.

This also saves 4k in binary space, as the macro apparently
duplicated a lot of code.