unit-test/keystore: simplify by removing granular mocking
cmocka mocking is a big pain. If you modify unit tests, it is very
hard to track down where and how the mocking has to be modified. The
unit test nearly becomes the implementation itself.

We can simplify and not mock all of this. The quality of the unit
tests is not impacted, as the mock_securechip.c already does KDF on
mock slot values and the resulting seed is checked, so if the
keystore.c implementation changes (e.g. using different KDF slots or
number of rounds or different hmac keys/values), the unit test will
fail.