miniscript: add Rust miniscript and bitcoin deps
rust-miniscript implements miniscrpit for the `wsh()` (Pay to witness
script hash) context as well as for the `tr()` context (Pay to
taproot). We use this dep as it works out of the box. The downside is
that this library uses a lot of generics and binary-size-inefficent
code. We could address much of that with PRs upstream, or in the worst
case roll our own miniscript implementation that is space-efficient in
the future.

The rust-bitcoin dep is needed explicitly as rust-miniscript requires
the `bitcoin::PublicKey` type to serialize pubkeys inside of
miniscript expressions. Maybe we can make upstream changes that allow
us to use our own types for that, as we already have all the code
needed to serialize pubkeys. rust-bitcoin is a dependency to
rust-miniscript as rust-bitcoin contains the necessary script
generation functions.

secp256k1 and secp256k1-sys are also transitive dependencies, as
`bitcoin::PublicKey` is a re-export of `secp256k1::PublicKey`. Also
here, we might be able to make upstream changes to remove the reliance
on these deps.

bech32 is updated to 0.9 as that is used by rust-bitcoin, so we don't
vendor two versions of the same lib.