miniscript: check for duplicate keys in policy
In a pubkey reference like `@0/<10;11>/*`, the receive address
derivation is `<xpub>/10/<addressIndex>` and the change deriation is
`<xpub>/11/<addressIndex>`. The pubkey reference would be replaced by
a pubkey derived at a particular change/receive derivation.

We check that a policy cannot contain the same pubkeys under any
derivation because Miniscript prohibits duplicate keys.