miniscript: derive witness script and address at keypath
Given a full keypath, we want to derive the witness script there to be
able to 1) sign at that keypath 2) show an address at that keypath.

The keypath is a different derivation mechanism to deriving using the
"native" derivation mechnism of providing the (is_change,address_index) tuple.

Example: wsh(and_v(v:pk(@0/<10;11>/*),pk(@1/<20;21>/*))) with our key [fp/48'/1'/0'/3']xpub...]
derived using keypath m/48'/1'/0'/3'/11/5 derives:
wsh(and_v(v:pk(@0/11/5),pk(@1/21/5))).

For signing at a keypath, we need to be able derive using the keypath,
as PSBTs etc. all have a keypath at the input to be signed.

For showing receive addresses, it is less clear if the API should take
a full keypath or a (is_change,address_index) tuple. We'll go with the
full keypath option for now as it is more in line with the other
script types (single sig, multisig).